requirements.txt
gpiozero==2.0.1
lgpio==0.2.2.0
openai==1.30.5
pillow==10.3.0
python-dotenv==1.0.1
requests==2.32.3
rpi-gpio==0.7.1
spidev==3.6
